+++ src/ts/AdUnits/VPAIDAdUnit.ts
@@ -100,12 +100,13 @@ export class VPAIDAdUnit extends AbstractAdUnit {
         this.setFinishState(FinishState.ERROR);
         Diagnostics.trigger('vpaid_load_timeout', new DiagnosticError(new Error('VPAID failed to load within timeout'), {
             id: this._vpaidCampaign.getId()
-        }));
+        }), this._vpaidCampaign.getSession());
         this.hide();
     }
 
     private onShow() {
         this.setShowing(true);
+        this.onStart.trigger();
         this._timer.start();
 
         if (this._nativeBridge.getPlatform() === Platform.IOS) {
@@ -154,7 +155,6 @@ export class VPAIDAdUnit extends AbstractAdUnit {
         this._timer.stop();
         this._view.updateTimeoutWidget();
         this._view.showAd();
-        this._nativeBridge.Listener.sendStartEvent(this._placement.getId());
     }
 
     private onAdError() {
@@ -181,7 +181,7 @@ export class VPAIDAdUnit extends AbstractAdUnit {
     private onAdStuck() {
         Diagnostics.trigger('vpaid_ad_stuck', new DiagnosticError(new Error('Ad playback stuck'), {
             campaignId: this._campaign.getId()
-        }));
+        }), this._campaign.getSession());
         this.setFinishState(FinishState.ERROR);
         this.hide();
     }
@@ -324,7 +324,7 @@ export class VPAIDAdUnit extends AbstractAdUnit {
         const impressionUrls = this._vpaidCampaign.getImpressionUrls();
         if (impressionUrls) {
             for (const impressionUrl of impressionUrls) {
-                this.sendThirdPartyEvent('vast impression', impressionUrl);
+                this.sendThirdPartyEvent('vpaid impression', impressionUrl);
             }
         }
     }