+++ src/ts/Views/PlayableMRAID.ts
@@ -160,7 +160,7 @@ export class PlayableMRAID extends MRAIDView {
                     const resourceUrl = this._campaign.getResourceUrl();
                     Diagnostics.trigger('playable_prepare_timeout', {
                         'url': resourceUrl ? resourceUrl.getOriginalUrl() : ''
-                    });
+                    }, this._campaign.getSession());
 
                     this._prepareTimeout = undefined;
                 }, 4500);