+++ src/ts/WebView.ts
@@ -37,6 +37,7 @@ import { AnalyticsStorage } from 'Analytics/AnalyticsStorage';
 import { StorageType } from 'Native/Api/Storage';
 import { FocusManager } from 'Managers/FocusManager';
 import { OperativeEventManager } from 'Managers/OperativeEventManager';
+import { SdkStats } from 'Utilities/SdkStats';
 
 import CreativeUrlConfiguration from 'json/CreativeUrlConfiguration.json';
 import CreativeUrlResponseAndroid from 'json/CreativeUrlResponseAndroid.json';
@@ -108,6 +109,7 @@ export class WebView {
 
             HttpKafka.setRequest(this._request);
             HttpKafka.setClientInfo(this._clientInfo);
+            SdkStats.setInitTimestamp();
 
             return this._deviceInfo.fetch();
         }).then(() => {
@@ -143,6 +145,7 @@ export class WebView {
                 this._focusManager.setListenAppBackground(true);
             } else {
                 this._focusManager.setListenScreen(true);
+                this._focusManager.setListenAndroidLifecycle(true);
             }
 
             return this.setupTestEnvironment();
@@ -163,10 +166,6 @@ export class WebView {
             }
 
             if(this._configuration.isAnalyticsEnabled() || this._clientInfo.getGameId() === '14850' || this._clientInfo.getGameId() === '14851') {
-                if(this._nativeBridge.getPlatform() === Platform.ANDROID) {
-                    this._focusManager.setListenAndroidLifecycle(true);
-                }
-
                 this._analyticsManager = new AnalyticsManager(this._nativeBridge, this._wakeUpManager, this._request, this._clientInfo, this._deviceInfo, this._configuration, this._focusManager);
                 return this._analyticsManager.init().then(() => {
                     this._sessionManager.setGameSessionId(this._analyticsManager.getGameSessionId());
@@ -190,7 +189,10 @@ export class WebView {
 
             this._assetManager = new AssetManager(this._cache, this._configuration.getCacheMode(), this._deviceInfo);
             this._campaignManager = new CampaignManager(this._nativeBridge, this._configuration, this._assetManager, this._sessionManager, this._request, this._clientInfo, this._deviceInfo, this._metadataManager);
-            this._campaignRefreshManager = new CampaignRefreshManager(this._nativeBridge, this._wakeUpManager, this._campaignManager, this._configuration);
+            this._campaignRefreshManager = new CampaignRefreshManager(this._nativeBridge, this._wakeUpManager, this._campaignManager, this._configuration, this._focusManager);
+
+            SdkStats.initialize(this._nativeBridge, this._request, this._configuration, this._sessionManager, this._campaignManager, this._metadataManager);
+
             return this._campaignRefreshManager.refresh();
         }).then(() => {
             this._wakeUpManager.onNetworkConnected.subscribe(() => this.onNetworkConnected());
@@ -247,6 +249,8 @@ export class WebView {
             return;
         }
 
+        SdkStats.sendShowEvent(placementId);
+
         if(campaign.isExpired()) {
             this.showError(true, placementId, 'Campaign has expired');
             this._campaignRefreshManager.refresh();
@@ -255,7 +259,7 @@ export class WebView {
                 id: campaign.getId(),
                 willExpireAt: campaign.getWillExpireAt()
             });
-            Diagnostics.trigger('campaign_expired', error);
+            Diagnostics.trigger('campaign_expired', error, campaign.getSession());
             return;
         }
 
@@ -281,7 +285,7 @@ export class WebView {
                 const error = new DiagnosticError(new Error('No connection is available'), {
                     id: campaign.getId(),
                 });
-                Diagnostics.trigger('mraid_no_connection', error);
+                Diagnostics.trigger('mraid_no_connection', error, campaign.getSession());
                 return;
             }